usize -> int + lisp(progress, retry, select) -> rust

+++ godot-simulation-client/src/serde.rs
@@ -255,9 +255,9 @@ pub fn parse_into_lvalue(se: &SExpr) -> Result<LValueS, ()> {
     match se {
         SExpr::Atom(atom) => {
             //Test if its an int
-            return match atom.canonical_str().parse::<i64>() {
+            return match atom.canonical_str().parse::<i32>() {
                 Ok(int) => Ok(LValueS::Int(int)),
-                Err(_) => match atom.canonical_str().parse::<f64>() {
+                Err(_) => match atom.canonical_str().parse::<f32>() {
                     //Test if its a float
                     Ok(float) => Ok(LValueS::Float(float)),
                     Err(_) => match atom.canonical_str() {
